Combine {Edit,New}MeetingModal into one component
