+++ app/components/meeting/edit-meeting-modal.hbs
@@ -1,61 +1,131 @@
 <Auk::Modal @size="large">
-  <Auk::Modal::Header @title={{t "edit-meeting"}} @onClose={{@onCancel}} />
+  <Auk::Modal::Header
+    @title={{if @meeting.isNew (t "new-agenda") (t "edit-meeting")}}
+    @closeAble={{true}}
+    @onClose={{@onCancel}}
+  />
   <Auk::Modal::Body>
     <Utils::KindSelector
-      @selectedKind={{this.kind}}
       @onChange={{this.setKind}}
+      @selectedKind={{this.kind}}
       @isLoading={{this.initializeKind.isRunning}}
     />
-    <div class="auk-u-mb-2">
-      <div class="auk-form-group">
-        <Auk::Label>{{t "start-date"}}</Auk::Label>
-        <WebComponents::VlDatepicker
-          @enableTime={{true}}
-          @date={{array this.startDate}}
-          @dateChanged={{this.selectStartDate}}
-        />
+    {{#if this.isAnnexMeeting}}
+      <div class="auk-u-mb-2">
+        <div class="auk-form-group">
+          <Auk::Label>
+            {{t "related-main-meeting-field-label"}}
+          </Auk::Label>
+          <Utils::ModelSelector
+            data-test-edit-meeting-related-main-meeting
+            @modelName="meeting"
+            @searchField="number-representation"
+            @sortField="-planned-start"
+            @selectedItems={{this.selectedMainMeeting}}
+            @selectModel={{this.selectMainMeeting}}
+            @isLoading={{this.initializeMainMeeting.isRunning}}
+            as |model|
+          >
+            {{kind-print-label model.kind}}
+            {{t "of"}}
+            {{moment-format model.plannedStart "DD-MM-YYYY"}}
+            ({{model.numberRepresentation}})
+          </Utils::ModelSelector>
+        </div>
       </div>
-    </div>
-    <div class="auk-u-mb-2">
-      <div class="auk-form-group">
-        <Auk::Label>{{t "meeting-number"}}</Auk::Label>
-        <Auk::Input
-          @type="number"
-          @block={{true}}
-          @value={{this.meetingNumber}}
-          data-test-edit-meeting-meeting-number
-        />
+    {{else}}
+      <div class="auk-u-mb-2">
+        <div class="auk-form-group" data-test-edit-meeting-datepicker>
+          <Auk::Label>{{t "start-date"}}</Auk::Label>
+          {{! TODO: Replace this with Auk::Datepicker, which isn't fully compliant yet (e.g. no time support) }}
+          <WebComponents::VlDatepicker
+            @enableTime={{true}}
+            @date={{array this.startDate}}
+            @dateChanged={{this.selectStartDate}}
+          />
+        </div>
       </div>
-    </div>
+    {{/if}}
     <div class="auk-u-mb-2">
       <div class="auk-form-group">
         <Auk::Label>
-          {{t "meeting-number-visual-representation"}}
+          {{t "meeting-number"}}
         </Auk::Label>
-        <Auk::Input
-          @block={{true}}
-          @value={{this.numberRepresentation}}
-          data-test-edit-meeting-number-representation
-        />
+        <div class="auk-o-flex">
+          {{#unless this.isAnnexMeeting}}
+            <Auk::Input
+              data-test-edit-meeting-meeting-number
+              @type="number"
+              @value={{this.meetingNumber}}
+              id="meetingNumberInput"
+              disabled={{this.initializeMeetingNumber.isRunning}}
+              @width="medium"
+              class="auk-u-mr-2"
+              required
+            />
+          {{/unless}}
+          <div class="auk-o-flex__item">
+            <div
+              data-test-edit-meeting-number-representation-view
+              class="auk-u-text-line-height--xsmall"
+            >
+              {{this.numberRepresentation}}
+            </div>
+            <Auk::ButtonLink
+              data-test-edit-meeting-number-representation-edit
+              disabled={{this.initializeMeetingNumber.isRunning}}
+              {{on "click" this.toggleEditingFormattedMeetingIdentifier}}
+            >
+              <small>{{t "edit"}}</small>
+            </Auk::ButtonLink>
+          </div>
+        </div>
       </div>
+      {{#if this.isEditingFormattedMeetingIdentifier}}
+        <div class="auk-o-flex">
+          <Auk::Input
+            @width="medium"
+            @value={{this.numberRepresentation}}
+            class="auk-u-mr-2"
+            data-test-edit-meeting-number-representation-input
+          />
+          <Auk::Button
+            @skin="primary"
+            {{on "click" this.toggleEditingFormattedMeetingIdentifier}}
+            data-test-edit-meeting-number-representation-save
+          >
+            {{t "save"}}
+          </Auk::Button>
+        </div>
+      {{/if}}
     </div>
-    <div class="auk-u-mb-2">
-      <div class="auk-form-group">
-        <Auk::Label>
-          {{t "meeting-location"}}
-        </Auk::Label>
-        <Auk::Input @block={{true}} @value={{this.extraInfo}} />
+    {{#unless this.isAnnexMeeting}}
+      <div class="auk-u-mb-2">
+        <div class="auk-form-group">
+          <Auk::Label>
+            {{t "meeting-location"}}
+          </Auk::Label>
+          <Auk::Input
+            @block={{true}}
+            @value={{this.extraInfo}}
+            data-test-edit-meeting-meeting-location
+          />
+        </div>
       </div>
-    </div>
+    {{/unless}}
   </Auk::Modal::Body>
   <Auk::Modal::Footer @onCancel={{@onCancel}}>
     <Auk::Button
       @skin="primary"
-      {{on "click" (perform this.updateMeeting)}}
-      disabled={{or (this.initializeKind.isRunning this.updateMeeting.isRunning)}}
+      {{on "click" (perform this.saveMeeting)}}
+      disabled={{this.savingIsDisabled}}
       data-test-edit-meeting-save
     >
-      {{t "edit"}}
+      {{#if @meeting.isNew}}
+        {{t "add"}}
+      {{else}}
+        {{t "edit"}}
+      {{/if}}
     </Auk::Button>
   </Auk::Modal::Footer>
 </Auk::Modal>
