--- app/components/meeting/new-meeting-modal.hbs
@@ -1,121 +1,0 @@
-<Auk::Modal>
-  <Auk::Modal::Header
-    @title={{t "new-agenda"}}
-    @closeAble={{true}}
-    @onClose={{@onCancel}}
-  />
-  <Auk::Modal::Body>
-    <Utils::KindSelector
-      data-test-new-meeting-kind
-      @onChange={{this.setKind}}
-      @selectedKind={{this.kind}}
-      @isLoading={{this.initializeKind.isRunning}}
-    />
-    {{#if this.isAnnexMeeting}}
-      <div class="auk-u-mb-2">
-        <div class="auk-form-group">
-          <Auk::Label>{{t "related-main-meeting-field-label"}}</Auk::Label>
-          <Utils::ModelSelector
-            data-test-new-meeting-related-main-meeting
-            @modelName="meeting"
-            @searchField="number-representation"
-            @sortField="-planned-start"
-            @selectedItems={{this.selectedMainMeeting}}
-            @selectModel={{this.selectMainMeeting}}
-            as |model|
-          >
-            {{kind-print-label model.kind}}
-            {{t "of"}}
-            {{moment-format model.plannedStart "DD-MM-YYYY"}}
-            ({{model.numberRepresentation}})
-          </Utils::ModelSelector>
-        </div>
-      </div>
-    {{else}}
-      <div class="auk-u-mb-2">
-        <div class="auk-form-group" data-test-new-meeting-datepicker>
-          <Auk::Label>{{t "start-date"}}</Auk::Label>
-          {{! TODO: Replace this with Auk::Datepicker, which isn't fully compliant yet (e.g. no time support) }}
-          <WebComponents::VlDatepicker
-            @enableTime={{true}}
-            @defaultHour={{10}}
-            @defaultMinute={{0}}
-            @dateChanged={{this.selectStartDate}}
-          />
-        </div>
-      </div>
-    {{/if}}
-    <div class="auk-u-mb-2">
-      <div class="auk-form-group">
-        <Auk::Label for="meetingNumberInput">{{t "meeting-number"}}</Auk::Label>
-        <div class="auk-o-flex">
-          {{#unless this.isAnnexMeeting}}
-            <Auk::Input
-              data-test-new-meeting-meeting-number
-              @type="number"
-              @value={{this.meetingNumber}}
-              id="meetingNumberInput"
-              disabled={{this.initializeMeetingNumber.isRunning}}
-              @width="medium"
-              class="auk-u-mr-2"
-            />
-          {{/unless}}
-          <div class="auk-o-flex__item">
-            <div
-              data-test-new-meeting-number-representation-view
-              class="auk-u-text-line-height--xsmall"
-            >
-              {{this.formattedMeetingIdentifier}}
-            </div>
-            <Auk::ButtonLink
-              data-test-new-meeting-number-representation-edit
-              disabled={{this.initializeMeetingNumber.isRunning}}
-              {{on "click" this.toggleEditingFormattedMeetingIdentifier}}
-            >
-              <small>{{t "edit"}}</small>
-            </Auk::ButtonLink>
-          </div>
-        </div>
-      </div>
-      {{#if this.isEditingFormattedMeetingIdentifier}}
-        <div class="auk-o-flex">
-          <Auk::Input
-            @width="medium"
-            @value={{this.formattedMeetingIdentifier}}
-            class="auk-u-mr-2"
-            data-test-new-meeting-number-representation-input
-          />
-          <Auk::Button
-            @skin="primary"
-            {{on "click" this.toggleEditingFormattedMeetingIdentifier}}
-            data-test-new-meeting-number-representation-save
-          >
-            {{t "save"}}
-          </Auk::Button>
-        </div>
-      {{/if}}
-    </div>
-    {{#unless this.isAnnexMeeting}}
-      <div class="auk-u-mb-2">
-        <div class="auk-form-group">
-          <Auk::Label>{{t "meeting-location"}}</Auk::Label>
-          <Auk::Input
-            @block={{true}}
-            @value={{this.extraInfo}}
-            data-test-new-meeting-meeting-location
-          />
-        </div>
-      </div>
-    {{/unless}}
-  </Auk::Modal::Body>
-  <Auk::Modal::Footer @onCancel={{@onCancel}}>
-    <Auk::Button
-      @skin="primary"
-      {{on "click" (perform this.createMeeting)}}
-      disabled={{this.createMeeting.isRunning}}
-      data-test-new-meeting-save
-    >
-      {{t "add"}}
-    </Auk::Button>
-  </Auk::Modal::Footer>
-</Auk::Modal>
